Acabada l'entrega 4.1 i preparada per fer pull-request
